Created statsservice to handle updating stats when collection or plant is added, updated or deleted Upon logging out we no longer remove the entire connection, just the email, the connection is removed on onclose
