+++ Shared/Models/Stats.cs
@@ -1,0 +1,8 @@
+﻿namespace Shared.Models;
+
+public class Stats
+{
+    public int TotalPlants { get; set; }
+    public int HappyPlants { get; set; }
+    public int Collections { get; set; }
+}
