+++ api/Events/PlantEvents/Client/ClientWantsToCreatePlant.cs
@@ -1,6 +1,7 @@
 ﻿using api.Core.Services;
 using api.EventFilters;
 using api.Events.PlantEvents.Server;
+using api.Events.Statistics;
 using api.Extensions;
 using Fleck;
 using lib;
@@ -15,12 +16,13 @@ public class ClientWantsToCreatePlantDto: BaseDtoWithJwt
 }
 
 [ValidateDataAnnotations]
-public class ClientWantsToCreatePlant(PlantService plantService, JwtService jwtService): BaseEventHandler<ClientWantsToCreatePlantDto>
+public class ClientWantsToCreatePlant(PlantService plantService, JwtService jwtService, StatsService statsService): BaseEventHandler<ClientWantsToCreatePlantDto>
 {
     public override async Task Handle(ClientWantsToCreatePlantDto dto, IWebSocketConnection socket)
     {
         var email = jwtService.GetEmailFromJwt(dto.Jwt!);
         var plant = await plantService.CreatePlant(dto.CreatePlantDto, email);
+        var stats = await statsService.GetStats(email);
         
         var serverCreatesNewPlant = new ServerSavesPlant
         {
@@ -28,5 +30,7 @@ public class ClientWantsToCreatePlant(PlantService plantService, JwtService jwtS
         };
         
         socket.SendDto(serverCreatesNewPlant);
+        
+        socket.SendDto(new ServerSendsStats{Stats = stats});
     }
 }
