+++ api/Events/PlantEvents/Client/ClientWantsToDeletePlant.cs
@@ -3,6 +3,7 @@ using api.Core.Services;
 using api.EventFilters;
 using api.Events.Global;
 using api.Events.PlantEvents.Server;
+using api.Events.Statistics;
 using api.Extensions;
 using Fleck;
 using lib;
@@ -16,13 +17,17 @@ public class ClientWantsToDeletePlantDto: BaseDtoWithJwt
 }
 
 [ValidateDataAnnotations]
-public class ClientWantsToDeletePlant(PlantService plantService, JwtService jwtService): BaseEventHandler<ClientWantsToDeletePlantDto>
+public class ClientWantsToDeletePlant(PlantService plantService, JwtService jwtService, StatsService statsService): BaseEventHandler<ClientWantsToDeletePlantDto>
 {
     public override async Task Handle(ClientWantsToDeletePlantDto dto, IWebSocketConnection socket)
     {
         var email = jwtService.GetEmailFromJwt(dto.Jwt!);
+        var stats = await statsService.GetStats(email);
+        
         await plantService.DeletePlant(dto.PlantId, email);
         socket.SendDto( new ServerConfirmsDelete());
+        
+        socket.SendDto(new ServerSendsStats{Stats = stats});
     }
 }
