+++ api/Events/PlantEvents/Client/ClientWantsToUpdatePlant.cs
@@ -1,5 +1,6 @@
 ﻿using api.Core.Services;
 using api.Events.PlantEvents.Server;
+using api.Events.Statistics;
 using api.Extensions;
 using Fleck;
 using lib;
@@ -13,15 +14,19 @@ public class ClientWantsToUpdatePlantDto: BaseDtoWithJwt
     public required UpdatePlantDto UpdatePlantDto { get; set; }
 }
 
-public class ClientWantsToUpdatePlant(PlantService plantService, JwtService jwtService): BaseEventHandler<ClientWantsToUpdatePlantDto>
+public class ClientWantsToUpdatePlant(PlantService plantService, JwtService jwtService, StatsService statsService): BaseEventHandler<ClientWantsToUpdatePlantDto>
 {
     public override async Task Handle(ClientWantsToUpdatePlantDto dto, IWebSocketConnection socket)
     {
         var email = jwtService.GetEmailFromJwt(dto.Jwt!);
         var plant = await plantService.UpdatePlant(dto.UpdatePlantDto, email);
+        var stats = await statsService.GetStats(email);
+        
         socket.SendDto(new ServerSavesPlant
         {
             Plant = plant
         });
+        
+        socket.SendDto(new ServerSendsStats{Stats = stats});
     }
 }
