+++ api/WebSocketConnectionService.cs
@@ -19,6 +19,12 @@ public class WebSocketConnectionService
         var clientId = connection.ConnectionInfo.Id;
         _connectedClients[clientId].Email = email;
     }
+    
+    public void RemoveEmailFromConnection(IWebSocketConnection connection)
+    {
+        var clientId = connection.ConnectionInfo.Id;
+        _connectedClients[clientId].Email = null;
+    }
 
     public void RemoveConnection(IWebSocketConnection connection)
     {
